+++ src/main/kotlin/me/lian/sts/api/SuspendingSTS.kt
@@ -1,7 +1,15 @@
 package me.lian.sts.api
 
 import kotlinx.coroutines.*
-import me.lian.sts.data.*
+import me.lian.sts.data.event.Event
+import me.lian.sts.data.event.EventType
+import me.lian.sts.data.facility.FacilityLayout
+import me.lian.sts.data.platform.Platform
+import me.lian.sts.data.system.SystemInformation
+import me.lian.sts.data.time.Time
+import me.lian.sts.data.timetable.Timetable
+import me.lian.sts.data.traindetails.TrainDetails
+import me.lian.sts.data.trainlist.Train
 import me.lian.sts.exception.*
 
 /**
