--- src/main/kotlin/me/lian/sts/data/Data.kt
@@ -1,11 +1,0 @@
-package me.lian.sts.data
-
-/**
- * A request to the plugin api.
- */
-sealed interface DataRequest
-
-/**
- * A response from the plugin api.
- */
-sealed interface DataResponse
