+++ src/main/kotlin/me/lian/sts/data/DataRequest.kt
@@ -1,0 +1,6 @@
+package me.lian.sts.data
+
+/**
+ * A request to the plugin api.
+ */
+interface DataRequest
