+++ src/main/kotlin/me/lian/sts/data/DataResponse.kt
@@ -1,0 +1,6 @@
+package me.lian.sts.data
+
+/**
+ * A response from the plugin api.
+ */
+interface DataResponse
