+++ src/main/kotlin/me/lian/sts/data/trainlist/Train.kt
@@ -1,24 +1,9 @@
-package me.lian.sts.data
+package me.lian.sts.data.trainlist
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import me.lian.sts.data.DataResponse
 import nl.adaptivity.xmlutil.serialization.XmlElement
-
-/**
- * Request to get a list of all trains.
- */
-@Serializable
-@SerialName("zugliste")
-internal data object TrainListRequest : DataRequest
-
-/**
- * A list of all trains.
- *
- * @property trains The list of trains.
- */
-@Serializable
-@SerialName("zugliste")
-data class TrainList(val trains: List<Train>) : DataResponse
 
 /**
  * A train.
